Move files into place
