+++ intern.json
@@ -48,7 +48,6 @@
 			"_tests/tests/unit/lib/middleware/**/*.js",
 			"_tests/tests/unit/lib/node/ErrorFormatter.js",
 			"_tests/tests/unit/lib/node/util.js",
-			"_tests/tests/unit/lib/reporters/BaseCoverage.js",
 			"_tests/tests/unit/lib/reporters/Benchmark.js",
 			"_tests/tests/unit/lib/reporters/Cobertura.js",
 			"_tests/tests/unit/lib/reporters/Coverage.js",
@@ -61,6 +60,7 @@
 			"_tests/tests/unit/lib/reporters/Runner.js",
 			"_tests/tests/unit/lib/reporters/Simple.js",
 			"_tests/tests/unit/lib/reporters/TeamCity.js",
+			"_tests/tests/unit/lib/reporters/TextCoverage.js",
 			"_tests/tests/unit/tasks/intern.js"
 		],
 		"plugins": [
