+++ tests/tsconfig.json
@@ -24,7 +24,6 @@
 		"./unit/lib/interfaces/tdd.ts",
 		"./unit/lib/middleware/*.ts",
 		"./unit/lib/node/*.ts",
-		"./unit/lib/reporters/BaseCoverage.ts",
 		"./unit/lib/reporters/Benchmark.ts",
 		"./unit/lib/reporters/Cobertura.ts",
 		"./unit/lib/reporters/Console.ts",
@@ -40,6 +39,7 @@
 		"./unit/lib/reporters/Runner.ts",
 		"./unit/lib/reporters/Simple.ts",
 		"./unit/lib/reporters/TeamCity.ts",
+		"./unit/lib/reporters/TextCoverage.ts",
 		"./unit/lib/*.ts",
 		"./unit/loaders/*.ts",
 		"./unit/tasks/*.ts",
