+++ tests/unit/lib/reporters/Coverage.ts
@@ -1,23 +1,157 @@
-import Coverage from 'src/lib/reporters/Coverage';
+import { CoverageMap } from 'istanbul-lib-coverage';
+import { spy, stub } from 'sinon';
+import _BaseCoverage, {
+	BaseCoverageOptions
+} from 'src/lib/reporters/BaseCoverage';
 
 const { registerSuite } = intern.getPlugin('interface.object');
 const { assert } = intern.getPlugin('chai');
+const mockRequire = intern.getPlugin<mocking.MockRequire>('mockRequire');
 
-registerSuite('lib/reporters/Coverage', {
-	construct() {
-		const reporter = new Coverage(<any>{ on() {} });
-		assert.equal(reporter.reportType, 'text');
-	},
+interface FullCoverage extends _BaseCoverage {
+	new (executor: Node, options: BaseCoverageOptions): _BaseCoverage;
+}
 
-	'#getReporterOptions': {
-		'maxColumns included'() {
-			const reporter = new Coverage(<any>{ on() {} }, { maxColumns: 80 });
-			assert.equal(reporter.getReporterOptions().maxColumns, 80);
+registerSuite('lib/reporters/BaseCoverage', function() {
+	const mockExecutor = <any>{
+		formatError: spy(),
+		on: spy(),
+		sourceMapStore: {
+			transformCoverage: spy(() => {
+				return { map: {} };
+			})
+		}
+	};
+
+	const mockGlobal: { [name: string]: any } = {};
+	const mockVisit = spy();
+	const mockSummarizers = {
+		pkg: spy(() => {
+			return { visit: mockVisit };
+		})
+	};
+	const mockCreate = spy();
+	const mockCreateCoverageMap = stub().returns({});
+
+	let Coverage: FullCoverage;
+	let removeMocks: () => void;
+
+	return {
+		before() {
+			return mockRequire(require, 'src/lib/reporters/BaseCoverage', {
+				'@dojo/shim/global': { default: mockGlobal },
+				'istanbul-lib-coverage': {
+					createCoverageMap: mockCreateCoverageMap
+				},
+				'istanbul-lib-report': {
+					createContext() {
+						return {};
+					},
+					summarizers: mockSummarizers
+				},
+				'istanbul-reports': { create: mockCreate }
+			}).then(handle => {
+				removeMocks = handle.remove;
+				Coverage = <any> class extends handle.module.default { };
+			});
 		},
 
-		'maxColumns not included'() {
-			const reporter = new Coverage(<any>{ on() {} }, {});
-			assert.equal(reporter.getReporterOptions().maxColumns, undefined);
+		after() {
+			removeMocks();
+		},
+
+		beforeEach() {
+			mockExecutor.formatError.reset();
+			mockExecutor.on.reset();
+			mockCreate.reset();
+			mockVisit.reset();
+			mockCreateCoverageMap.reset();
+		},
+
+		tests: {
+			construct() {
+				const watermarks = <any>{};
+				const reporter = new Coverage(mockExecutor, <any>{
+					filename: 'foo',
+					watermarks
+				});
+				assert.propertyVal(reporter, 'filename', 'foo');
+				assert.propertyVal(reporter, 'watermarks', watermarks);
+			},
+
+			'#createCoverageReport': {
+				'without data'() {
+					mockExecutor.coverageMap = { files: () => [] };
+					const reporter = new Coverage(mockExecutor, <any>{});
+					reporter.createCoverageReport('text', {});
+					assert.equal(mockVisit.callCount, 1);
+					assert.equal(mockCreateCoverageMap.callCount, 1);
+					assert.equal(mockCreate.callCount, 1);
+					assert.equal(
+						mockCreate.getCall(0).args[0],
+						'text',
+						'report should be text by default'
+					);
+				},
+
+				'with data'() {
+					mockExecutor.coverageMap = { files: () => [] };
+					const reporter = new Coverage(mockExecutor, <any>{});
+					reporter.createCoverageReport('json', <CoverageMap>{
+						files() {}
+					});
+					assert.equal(mockVisit.callCount, 1);
+					assert.equal(mockCreateCoverageMap.callCount, 0);
+					assert.equal(mockCreate.callCount, 1);
+					assert.equal(
+						mockCreate.getCall(0).args[0],
+						'json',
+						'report should be assigned value'
+					);
+				}
+			},
+
+			'#getReporterOptions': {
+				'filename included'() {
+					const reporter = new Coverage(<any>{ on() {} }, {
+						filename: 'foo'
+					});
+					assert.deepEqual(reporter.getReporterOptions(), {
+						file: 'foo'
+					});
+				},
+
+				'filename not included'() {
+					const reporter = new Coverage(<any>{ on() {} }, {});
+					assert.deepEqual(reporter.getReporterOptions(), {
+						file: undefined
+					});
+				}
+			},
+
+			'#runEnd': {
+				'without data'() {
+					const reporter = new Coverage(mockExecutor, <any>{});
+					const create = stub(reporter, 'createCoverageReport');
+					mockExecutor.coverageMap = { files: () => [] };
+					reporter.runEnd();
+					assert.equal(create.callCount, 0);
+				},
+
+				'with data'() {
+					mockExecutor.coverageMap = {};
+					const reporter = new Coverage(mockExecutor, <any>{});
+					const create = stub(reporter, 'createCoverageReport');
+					mockExecutor.coverageMap = { files: () => ['foo.js'] };
+					reporter.runEnd();
+					assert.equal(create.callCount, 1);
+					assert.equal(create.getCall(0).args[0], undefined);
+					assert.equal(
+						create.getCall(0).args[1],
+						mockExecutor.coverageMap
+					);
+				}
+			}
 		}
-	}
+	};
 });
